// dsl/packages/cli/test/test-specs/gen-out/client/typescript-client/src/model/EnumRecord.ts
// Generated by RSD - Do not modify
import { isSampleEnum, type SampleEnum } from './SampleEnum.js';
import { checkOptProp, checkProp, createTypedArrayGuard, isNull, isRecord, propListValue, propValue } from '../_type-utils.js';

export type EnumRecord = {
	readonly value: SampleEnum;
	readonly value_Null: SampleEnum | null;
	readonly value_Opt?: SampleEnum;
	readonly value_Opt_Null?: SampleEnum | null;
	readonly list: SampleEnum[];
	readonly list_Null: SampleEnum[] | null;
	readonly list_Opt?: SampleEnum[];
	readonly list_Opt_Null?: SampleEnum[] | null;
};

export function isEnumRecord(value: unknown): value is EnumRecord {
	return isRecord(value) &&
		checkProp(value, 'value', isSampleEnum) &&
		(checkProp(value, 'value_Null', isNull) || checkProp(value, 'value_Null', isSampleEnum)) &&
		checkOptProp(value, 'value_Opt', isSampleEnum) &&
		(checkOptProp(value, 'value_Opt_Null', isNull) || checkOptProp(value, 'value_Opt_Null', isSampleEnum)) &&
		checkProp(value, 'list', createTypedArrayGuard(isSampleEnum)) &&
		(checkProp(value, 'list_Null', isNull) || checkProp(value, 'list_Null', createTypedArrayGuard(isSampleEnum))) &&
		checkOptProp(value, 'list_Opt', createTypedArrayGuard(isSampleEnum)) &&
		(checkOptProp(value, 'list_Opt_Null', isNull) || checkOptProp(value, 'list_Opt_Null', createTypedArrayGuard(isSampleEnum)));
}

export function EnumRecordFromJSON($value: Record<string, unknown>): EnumRecord {
	const value = propValue('value', $value, isSampleEnum);
	const value_Null = propValue('value_Null', $value, isSampleEnum, 'null');
	const value_Opt = propValue('value_Opt', $value, isSampleEnum, 'optional');
	const value_Opt_Null = propValue('value_Opt_Null', $value, isSampleEnum, 'optional_null');
	const list = propListValue('list', $value, isSampleEnum);
	const list_Null = propListValue('list_Null', $value, isSampleEnum, 'null');
	const list_Opt = propListValue('list_Opt', $value, isSampleEnum, 'optional');
	const list_Opt_Null = propListValue('list_Opt_Null', $value, isSampleEnum, 'optional_null');
	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}

export function EnumRecordToJSON($value: EnumRecord): Record<string, unknown> {
	const value = $value.value;
	const value_Null = $value.value_Null;
	const value_Opt = $value.value_Opt;
	const value_Opt_Null = $value.value_Opt_Null;
	const list = $value.list;
	const list_Null = $value.list_Null;
	const list_Opt = $value.list_Opt;
	const list_Opt_Null = $value.list_Opt_Null;

	return {
		value,
		value_Null,
		value_Opt,
		value_Opt_Null,
		list,
		list_Null,
		list_Opt,
		list_Opt_Null,
	};
}